tools/requirements.txt
mkdocs-material
paho-mqtt
pyserial
requests
esptool
